refactor(guest-agent): extract cli stderr diagnostics
